+++ tiny_wizard/room/lab_dungeon_generator.gd
@@ -4,6 +4,7 @@ extends RefCounted
 
 const TUTORIAL_START_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_start_room.tscn")
 const START_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_formal_start_room.tscn")
+const TUTORIAL_TARGET_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_tutorial_target_room.tscn")
 const TUTORIAL_RAVEN_SAFEHOUSE_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_tutorial_raven_safehouse_room.tscn")
 const TUTORIAL_COMBAT_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_tutorial_combat_room.tscn")
 const TUTORIAL_BOSS_ROOM_SCENE := preload("res://tiny_wizard/room/room_types/lab_tutorial_boss_room.tscn")
@@ -109,18 +110,24 @@ const TUTORIAL_ROOM_LAYOUT := [
 	},
 	{
 		"coord": Vector2i(1, 0),
+		"type": "tutorial_targets",
+		"label": "Target Calibration Range",
+		"scene": TUTORIAL_TARGET_ROOM_SCENE,
+	},
+	{
+		"coord": Vector2i(2, 0),
 		"type": "tutorial_merchant",
 		"label": "Raven Tutorial Safehouse",
 		"scene": TUTORIAL_RAVEN_SAFEHOUSE_ROOM_SCENE,
 	},
 	{
-		"coord": Vector2i(2, 0),
+		"coord": Vector2i(3, 0),
 		"type": "tutorial_combat",
 		"label": "Breach Training Lab",
 		"scene": TUTORIAL_COMBAT_ROOM_SCENE,
 	},
 	{
-		"coord": Vector2i(3, 0),
+		"coord": Vector2i(4, 0),
 		"type": "tutorial_boss",
 		"label": "Low-Grade Fusion Chamber",
 		"scene": TUTORIAL_BOSS_ROOM_SCENE,